+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support for wasm32-unknown-unknown target.
 - Implement `ChokeItem` for `Result` and `Option`.
 - Allow updating the `ChokeStreamSettings` dynamically.
+- `ChokeSink` to wrap a `future::Sink`.
 
 ### Changed
 - Rename types.